Phase 2 Setup
just trying to implement the stuff the graders were looking for in TP1.
still need to update DatabaseHelper to match the new User attributes,
create pages for users to edit their own info, and create pages for
admins to edit other users' details along with the one-time password
feature. will continue working on this over the coming days